+++ test/conftest.py
@@ -10,7 +10,7 @@ def tmp_home(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> Path:
     monkeypatch.setenv("HOME", str(tmp_path))
     # At this point, why not always set the Windows var?
     monkeypatch.setenv("USERPROFILE", str(tmp_path))
-    # Used by appdirs and keyring:
+    # Used by platformdirs and keyring:
     monkeypatch.delenv("XDG_CONFIG_HOME", raising=False)
     monkeypatch.setenv("LOCALAPPDATA", str(tmp_path))
     return tmp_path